+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.5.2
+* Rename `ConfigManager` to `ConfigCatalog` (was missed in 0.5.1)
+* Rename `EquipmentFrame.performEquipmentAction` to `EquipmentFrame.performEquipmentActionServer` to improve transparency
+* Add `logbookModelParameters` and `SetLogbookModelParameters` to `EquipmentFrame` to mirror `ItemFrame`
+* Make 'Quick Add' overload for `SingleItemDisplayFrame.Add` implicitly support Equipment Drones
+* Add `GSUtil.AllChildren` extension for transforms
+* Fix `GSUtil.SetupItemDisplay` setting material instead of shared material
+* Add `UpdateItemFollowerScaleDebug` to attach to item followers when creating item displays (`ItemFollower` only sets scale once when instantiating the follower`
+* Fix `Configurable` attributes not applying to fields when within a module lacking a `Configurable` attribute itself
+
 ## 0.5.1
 * Add `ArtifactActionCatalog` to handle artifact actions on enable/disable
 * Add `enabledAction` and `disabledAction` to `ArtifactFrame`